Updated_Pipeline/requirements.txt
# LUNA PPE Safety Monitor - Requirements
# =========================================
# AI-powered PPE compliance monitoring system with YOLOv8, LLaVA, and RAG

# Core Framework
Flask==3.1.2
Werkzeug==3.1.3

# Computer Vision & Object Detection
opencv-python==4.12.0.88
ultralytics==8.3.228
torch==2.9.1
torchvision==0.24.1

# Image Captioning (LLaVA)
transformers==4.57.1
accelerate==1.12.0
bitsandbytes==0.48.2
safetensors==0.6.2
pillow==12.0.0

# RAG & Vector Database
chromadb==1.3.5
requests==2.32.5

# NLP & Language Models
tokenizers==0.22.1
huggingface-hub==0.36.0

# Data Processing
numpy==2.2.6
pandas==2.3.3
scipy==1.16.3

# Visualization
matplotlib==3.10.7

# Utilities
python-dateutil==2.9.0.post0
tqdm==4.67.1
coloredlogs==15.0.1
click==8.3.0
PyYAML==6.0.3

# Optional (already included in transformers/torch)
filelock==3.20.0
fsspec==2025.10.0
regex==2025.11.3
sympy==1.14.0
networkx==3.5
Jinja2==3.1.6

# Development Dependencies
python-dotenv==1.2.1
rich==14.2.0

# Note: Ollama must be installed separately
# Download from: https://ollama.ai
# Required models:
#   - ollama pull llama3
#   - ollama pull nomic-embed-text
